refactor: Cost 테스트 코드 파일 분리 및 테스트 추가
